+++ dxpApp/handelSrc/handel_xerxes.c
@@ -838,8 +838,8 @@ HANDEL_STATIC int xia__AddSystemFPGA(Module *module, char *sysFPGAName,
   sysFPGAStr[0] = (char *)handel_md_alloc(strlen(sysFPGAName) + 1);
 
   if (!sysFPGAStr[0]) {
-	sprintf(info_string, "Unable to allocate %d bytes for 'sysFPGAStr[0]'",
-			strlen(sysFPGAName) + 1);
+	sprintf(info_string, "Unable to allocate %ld bytes for 'sysFPGAStr[0]'",
+			(long)strlen(sysFPGAName) + 1);
 	xiaLogError("xia__AddSystemFPGA", info_string, XIA_NOMEM);
 	return XIA_NOMEM;
   }
@@ -1056,8 +1056,8 @@ HANDEL_STATIC int xia__AddSystemDSP(Module *module, char *sysDSPName,
   sysDSPStr[0] = (char *)handel_md_alloc(strlen(sysDSPName) + 1);
 
   if (!sysDSPStr[0]) {
-	sprintf(info_string, "Unable to allocate %d bytes for 'sysDSPStr[0]'",
-			strlen(sysDSPName) + 1);
+	sprintf(info_string, "Unable to allocate %ld bytes for 'sysDSPStr[0]'",
+			(long)strlen(sysDSPName) + 1);
 	xiaLogError("xiaAddSystemDSP", info_string, XIA_NOMEM);
 	return XIA_NOMEM;
   }
@@ -1181,8 +1181,8 @@ HANDEL_STATIC int xia__AddFiPPIA(Module *module, char *sysFippiAName,
   sysFippiAStr[0] = (char *)handel_md_alloc(strlen(sysFippiAName) + 1);
 
   if (!sysFippiAStr[0]) {
-	sprintf(info_string, "Unable to allocate %d bytes for 'sysFippiAStr[0]'",
-			strlen(sysFippiAName) + 1);
+	sprintf(info_string, "Unable to allocate %ld bytes for 'sysFippiAStr[0]'",
+			(long)strlen(sysFippiAName) + 1);
 	xiaLogError("xiaAddSystemFippiA", info_string, XIA_NOMEM);
 	return XIA_NOMEM;
   }
@@ -1256,7 +1256,7 @@ HANDEL_STATIC int xia__AddXerxesBoardType(Module *m)
   type = (char **)handel_md_alloc(sizeof(char *));
   
   if (!type) {
-	sprintf(info_string, "Error allocating %d bytes for 'type'", sizeof(char *));
+	sprintf(info_string, "Error allocating %ld bytes for 'type'", (long)sizeof(char *));
 	xiaLogError("xia__AddXerxesBoardType", info_string, XIA_NOMEM);
 	return XIA_NOMEM;
   }
@@ -1266,8 +1266,8 @@ HANDEL_STATIC int xia__AddXerxesBoardType(Module *m)
   if (!type[0]) {
 	handel_md_free(type);
 	
-	sprintf(info_string, "Error allocating %d bytes for 'type[0]'",
-			strlen(m->type) + 1);
+	sprintf(info_string, "Error allocating %ld bytes for 'type[0]'",
+			(long)strlen(m->type) + 1);
 	xiaLogError("xia__AddXerxesBoardType", info_string, XIA_NOMEM);
 	return XIA_NOMEM;
   }
@@ -1388,8 +1388,8 @@ HANDEL_STATIC int xia__AddXerxesModule(Module *m)
 									sizeof(char *));
 
   if (!modStr) {
-	sprintf(info_string, "Error allocating %d bytes for 'modStr'",
-			(m->number_of_channels + 2) * sizeof(char *));
+	sprintf(info_string, "Error allocating %ld bytes for 'modStr'",
+			(long)((m->number_of_channels + 2) * sizeof(char *)));
 	xiaLogError("xia__AddXerxesModule", info_string, XIA_NOMEM);
 	return XIA_NOMEM;
   }
@@ -1513,8 +1513,8 @@ HANDEL_STATIC int xia__DoMMUConfig(Module *m)
 	mmu = (char **)handel_md_alloc(sizeof(char **));
 
 	if (!mmu) {
-	  sprintf(info_string, "Error allocating %d bytes for 'mmu'",
-			  sizeof(char **));
+	  sprintf(info_string, "Error allocating %ld bytes for 'mmu'",
+			  (long)sizeof(char **));
 	  xiaLogError("xia__DoMMUConfig", info_string, XIA_NOMEM);
 	  return XIA_NOMEM;
 	}
@@ -1524,8 +1524,8 @@ HANDEL_STATIC int xia__DoMMUConfig(Module *m)
 	if (!mmu[0]) {
 	  handel_md_free(mmu);
 
-	  sprintf(info_string, "Error allocating %d bytes for 'mmu[0]'",
-			  strlen(name) + 1);
+	  sprintf(info_string, "Error allocating %ld bytes for 'mmu[0]'",
+			  (long)strlen(name) + 1);
 	  xiaLogError("xia__DoMMUConfig", info_string, XIA_NOMEM);
 	  return XIA_NOMEM;
 	}
@@ -2120,8 +2120,8 @@ HANDEL_STATIC int xia__AddSystemFiPPI(Module *m, char *sysFipName,
   sysFipStr[0] = handel_md_alloc(strlen(sysFipName) + 1);
 
   if (sysFipStr[0] == NULL) {
-    sprintf(info_string, "Unable to allocated %d bytes for 'sysFipStr[0]'",
-            strlen(sysFipName) + 1);
+    sprintf(info_string, "Unable to allocated %ld bytes for 'sysFipStr[0]'",
+            (long)strlen(sysFipName) + 1);
     xiaLogError("xia__AddSystemFiPPI", info_string, XIA_NOMEM);
     return XIA_NOMEM;
   }
